AF_UNIX code looks good, lint happy, hangs in shuffle()
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@413 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/poker.d/payoff.c
@@ -6,12 +6,10 @@
 # include	"tellall.h"
 
 void	payoff( player, n_players, winner, pot )
-
 PLAYER	player[];		/* player information */
 int	n_players;		/* number of players */
 int	winner;			/* which player won */
-int	pot;			/* how much $ was won */
-
+long	pot;			/* how much $ was won */
 {
 char	temp[80];
